+++ CHANGES.rst
@@ -1,3 +1,42 @@
+CHANGELOG
+=========
+
+0.3.0
+-----
+-  Allow content to be filtered by language selection
+
+0.2.8
+-----
+-  Add support for flat pages
+
+0.2.7
+-----
+-  Add caching to `get_featured_category_pages`
+
+0.2.6
+-----
+-  Added sensible default for available_languages
+
+0.2.5
+-----
+-  Added support for translations
+
+0.2.4
+-----
+-  Allow top nav to be global variable
+
+0.2.2
+-----
+-  Use `utils.get_workspace()` to avoid duplication
+
+0.2.2
+-----
+-  Fix development.ini file
+
+0.2.1
+-----
+-  Bump required version for praekelt-python-gitmodel
+
 0.2
 ---
 -  Added `git.content_repo_url` for cloning when app starts